Broke down the codebase further into survey service to separate the project related logics from surveys logic.
