+++ mcp/server/tools/src/tools/ProjectTools.ts
@@ -14,18 +14,6 @@ export const projectTools: Tool[] = [
         }
       },
       {
-        name: "get_surveys_in_a_project",
-        description: "Fetch Voice Korea's Project's surveys by the ID of the project",
-        inputSchema: {
-          type: "object",
-          properties: {
-            question: { type: "string" },
-            id: { type: "number" }
-          },
-          required: ["question","id"]
-        }
-      },
-      {
         name: "search_projects_by_title",
         description: "Fetch Voice Korea's Projects by their title",
         inputSchema: {
